+++ gaios/LTFlow/Controllers/LTWithdraw/Cells/LTWithdrawNoteCell.swift
@@ -1,0 +1,22 @@
+import UIKit
+
+class LTWithdrawNoteCell: UITableViewCell {
+
+    @IBOutlet weak var bg: UIView!
+    @IBOutlet weak var textField: UITextField!
+
+    class var identifier: String { return String(describing: self) }
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        bg.cornerRadius = 5.0
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+    }
+
+    func configure(text: String) {
+        textField.text = text
+    }
+}