+++ gaios/Wallet/AccountViewController/AccountViewController.swift
@@ -804,6 +804,8 @@ extension AccountViewController: DialogScanViewControllerDelegate {
                 case .lnUrlAuth(let data):
                     // open LNURL-Auth page
                     ltAuthViewController(requestData: data)
+                case .lnUrlWithdraw(let data):
+                    ltWithdrawViewController(requestData: data)
                 default:
                     // open Send page
                     let tx = parser.createTx?.tx
@@ -824,6 +826,14 @@ extension AccountViewController: DialogScanViewControllerDelegate {
                     DropAlert().warning(message: error.localizedDescription)
                 }
             }
+        }
+    }
+
+    func ltWithdrawViewController(requestData: LnUrlWithdrawRequestData) {
+        let storyboard = UIStoryboard(name: "LTFlow", bundle: nil)
+        if let vc = storyboard.instantiateViewController(withIdentifier: "LTWithdrawViewController") as? LTWithdrawViewController {
+            vc.requestData = requestData
+            navigationController?.pushViewController(vc, animated: true)
         }
     }